Update StairDP question
A dp problem from cracking coding interview. Be careful with the bound
of the Integer Type in the language you use.
A child is running up a staircase with n steps, and can hop either 1
step, 2 steps or 3 steps at a time. Implement a method to count how many
possible ways the child can run up the stairs.